# docs/swagger.yaml
basePath: /
definitions:
...
  model.CreateOperatorRequest:
    properties:
      address:
        type: string
      city:
        maxLength: 100
        type: string
      company_name:
        maxLength: 255
        type: string
      contact_email:
        maxLength: 255
        type: string
      contact_person:
        maxLength: 255
        type: string
      contact_phone:
        maxLength: 20
        type: string
      country:
        maxLength: 100
        type: string
      established_at:
        type: integer
      license_expiry:
        type: integer
      license_number:
        maxLength: 100
        type: string
      notes:
        maxLength: 1000
        type: string
      operator_code:
        maxLength: 20
        type: string
      operator_type:
        maxLength: 100
        type: string
      postal_code:
        maxLength: 20
        type: string
      province:
        maxLength: 100
        type: string
      user_id:
        type: string
      website:
        maxLength: 500
        type: string
    required:
    - address
    - city
    - company_name
    - contact_email
    - contact_person
    - contact_phone
    - country
    - license_number
    - operator_code
    - operator_type
    - postal_code
    - province
    - user_id
    type: object
...
  model.PageMetadata:
    properties:
      current_page:
        type: integer
      from:
        type: integer
      last_page:
        type: integer
      per_page:
        type: integer
      to:
        type: integer
      total:
        type: integer
    type: object
...
  model.SwaggerPageResponse:
    properties:
      data:
        items: {}
        type: array
      errors:
        type: string
      message:
        type: string
      meta:
        $ref: '#/definitions/model.PageMetadata'
      success:
        type: boolean
    type: object
  model.SwaggerWebResponse:
    properties:
      data: {}
      errors:
        type: string
      message:
        type: string
      meta:
        $ref: '#/definitions/model.PageMetadata'
      success:
        type: boolean
    type: object
...
  model.UpdateOperatorRequest:
    properties:
      address:
        type: string
      city:
        maxLength: 100
        type: string
      company_name:
        maxLength: 255
        type: string
      contact_email:
        maxLength: 255
        type: string
      contact_person:
        maxLength: 255
        type: string
      contact_phone:
        maxLength: 20
        type: string
      country:
        maxLength: 100
        type: string
      established_at:
        type: integer
      is_active:
        type: boolean
      license_expiry:
        type: integer
      license_number:
        maxLength: 100
        type: string
      notes:
        maxLength: 1000
        type: string
      operator_code:
        maxLength: 20
        type: string
      operator_type:
        maxLength: 100
        type: string
      postal_code:
        maxLength: 20
        type: string
      province:
        maxLength: 100
        type: string
      status:
        enum:
        - active
        - inactive
        - suspended
        type: string
      website:
        maxLength: 500
        type: string
    type: object
...
host: localhost:3000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for Golang Clean Architecture.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: MKP Boarding Test API
  version: "1.0"
paths:
...
  /api/operators:
    get:
      consumes:
      - application/json
      description: Get list of operators with optional filtering
      parameters:
      - description: Filter by company name
        in: query
        name: company_name
        type: string
      - description: Filter by operator type
        in: query
        name: operator_type
        type: string
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Filter by country
        in: query
        name: country
        type: string
      - description: Filter by province
        in: query
        name: province
        type: string
      - description: Filter by city
        in: query
        name: city
        type: string
      - default: true
        description: Filter by active status
        in: query
        name: is_active
        type: boolean
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of operators
          schema:
            $ref: '#/definitions/model.SwaggerPageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
      security:
      - BearerAuth: []
      summary: List operators
      tags:
      - Operators
    post:
      consumes:
      - application/json
      description: Create a new operator with company information
      parameters:
      - description: Create operator request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.CreateOperatorRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Operator created successfully
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
      security:
      - BearerAuth: []
      summary: Create a new operator
      tags:
      - Operators
  /api/operators/{operatorId}:
    delete:
      consumes:
      - application/json
      description: Delete operator by operator ID
      parameters:
      - description: Operator ID
        in: path
        name: operatorId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Operator deleted successfully
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "404":
          description: Operator not found
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
      security:
      - BearerAuth: []
      summary: Delete operator
      tags:
      - Operators
    get:
      consumes:
      - application/json
      description: Get operator details by operator ID
      parameters:
      - description: Operator ID
        in: path
        name: operatorId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Operator details
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "404":
          description: Operator not found
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
      security:
      - BearerAuth: []
      summary: Get operator by ID
      tags:
      - Operators
    put:
      consumes:
      - application/json
      description: Update operator information by operator ID
      parameters:
      - description: Operator ID
        in: path
        name: operatorId
        required: true
        type: string
      - description: Update operator request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UpdateOperatorRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Operator updated successfully
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "404":
          description: Operator not found
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
      security:
      - BearerAuth: []
      summary: Update operator
      tags:
      - Operators
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"